Implement respect tool through save editing
Hey, first commit! Right, so -- this first commit does many things (indeed, even,
all the things). All the project bootstrap has been kinked out, rollup is
bundling stuff, and we have a barebones website that allows save editing.

The main JS file is mainly comprised of event handlers for the inputs and the
download buttons, and by the logic required to unzip, edit, and re-zip the save
files so that they can be parsed by Pathfinder Kingmaker.

Still untested, so this is very much a WIP so far.