Refactor logic that stores info from one save to the other
While previously we were explicitly storing the "global" alignment and experience
for all characters, expecting it to always be in the same place, this many times
resulted in errors being thrown to the users.

Instead, we now map all the ID's available in the save into an object and make use
of that to build a character data map that is stored and later used to transfer info
from one save to the other.

This could also serve as a baseline for a future character editor, since we already
get around the use of references in the save file.