Refactor application and implement per-character respec
Right, so -- changes in the commit are a bit too many to list,
but the jist is that the entire thing has been re-written in React
and has been re-organized and somewhat optimized.

Per-char respec has been implemented and is now available to select
through a list that is rendered after selecting the first save.

The required info from the first save is also being stored in the
local storage to avoid issues with page reloads.

A lot of logic related to keeping progress-related feats has been
written, but commented away since it's not currently working.

Linter has been implemented along with a pre-commit and pre-push hook.